refactor(dimoscope): standalone media node (:8092) — split camera WebRTC/WebCodecs out of the gateway
Gateway becomes a thin data-relay + the teleop/goal/RPC trust boundary; the camera
(encode-once-per-topic WebRTC + WebCodecs) moves to its own Zenoh-peer media node so it is
transport- and lifecycle-independent. Wire protocols unchanged; browser media adapters point at :8092.
